+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/dto/gamification/RankingResponseDto.java
@@ -1,11 +1,13 @@
 package com.itachallenge.challenge.dto.gamification;
 
 import lombok.AllArgsConstructor;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 
-@Data
+@Getter
+@Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class RankingResponseDto {