+++ itachallenge-challenge/src/main/java/com/itachallenge/gamification/service/PointsService.java
@@ -7,5 +7,5 @@ import reactor.core.publisher.Mono;
 
 public interface PointsService {
 
-    Flux<RankingResponseDto> getRankingAscOrder();
+    Flux<RankingResponseDto> getRankingDescOrder();
 }